+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
     - Issue 229 - Create new database tables for version d collections
+    - Issue 234 - Add documentation on collection versioning
 ### Changed
 ### Deprecated
 ### Removed